+++ C64/TAPContainer.h
@@ -35,7 +35,7 @@ private:
     uint8_t *data;
     
     //! @brief    File size
-    int size;
+    size_t size;
     
     /*! @brief    File pointer
      *  @details  An offset into the data array. 
@@ -71,8 +71,8 @@ public:
     const char *getTypeAsString() { return "TAP"; }
     
     bool fileIsValid(const char *filename);
-    bool readFromBuffer(const uint8_t *buffer, unsigned length);
-    unsigned writeToBuffer(uint8_t *buffer);
+    bool readFromBuffer(const uint8_t *buffer, size_t length);
+    size_t writeToBuffer(uint8_t *buffer);
     
     
     //
@@ -81,6 +81,6 @@ public:
     
     unsigned TAPversion() { return data[0x000C]; }
     uint8_t *getData() { return &data[0x0014]; }
-    int getSize() { return size - 0x14; }
+    size_t getSize() { return size - 0x14; }
 };
 #endif
